Implement common.Runnable for stats.Channel feature
